Shorten edit execute method
