+++ src/main/data/fleur.txt
@@ -1,3 +1,3 @@
-[D][ ] submit assignment (by: Feb 26 2025)
-[E][X] interview (from: Feb 25 2025 to: Feb 25 2025)
-[T][ ] woof
+[D][ ] moo (by: Feb 07 2025)
+[E][X] rawr (from: Feb 24 2026 to: Feb 27 2026)
+[T][ ] meow
