+++ src/main/java/fleur/commands/EditCommand.java
@@ -2,11 +2,7 @@ package fleur.commands;
 
 import fleur.tasks.TaskList;
 import fleur.tasks.Task;
-import fleur.tasks.ToDo;
-import fleur.tasks.Deadline;
-import fleur.tasks.Event;
 import fleur.exceptions.FleurException;
-import fleur.exceptions.FleurInvalidCommandException;
 
 public class EditCommand extends Command {
 
@@ -22,33 +18,12 @@ public class EditCommand extends Command {
         int index = Integer.parseInt(parts[1]) - 1;
         assert index >= 0 : "Task number must be greater than 0.";
         assert index <= tasks.size() : "Task number must be less than or equal to number of tasks";
-        Task oldTask = tasks.getTask(index);
-        String taskType = oldTask.getTaskType();
+        Task task = tasks.getTask(index);
 
         StringBuilder result = new StringBuilder();
         result.append("D'accord, your task has been edited:\n");
-
-        Task editedTask = null;
-        switch (taskType) {
-        case "todo":
-            ToDo oldToDo = (ToDo) oldTask;
-            EditToDoCommand editToDoCommand = new EditToDoCommand(description);
-            editedTask = editToDoCommand.editToDo(oldToDo);
-            break;
-        case "deadline":
-            Deadline oldDeadline = (Deadline) oldTask;
-            EditDeadlineCommand editDeadlineCommand = new EditDeadlineCommand(description);
-            editedTask = editDeadlineCommand.editDeadline(oldDeadline);
-            break;
-        case "event":
-            Event oldEvent = (Event) oldTask;
-            EditEventCommand editEventCommand = new EditEventCommand(description);
-            editedTask = editEventCommand.editEvent(oldEvent);
-            break;
-        default:
-            throw new FleurInvalidCommandException();
-        }
-        result.append(editedTask);
+        task.edit(description);
+        result.append(task);
         return result.toString();
     }
 }
