+++ src/main/java/fleur/tasks/ToDo.java
@@ -15,10 +15,6 @@ public class ToDo extends Task {
         return "todo";
     }
 
-    public void edit(String newDescription) {
-        this.description = newDescription;
-    }
-
     @Override
     public String toString() {
         return "[T]" + super.toString();
